fix(ci): avoid Linux clipboard build dependency
